requirements.txt
requests
pillow
imageio
opencv-python
imageio-ffmpeg
